requirements.txt
pyodbc
python-dotenv
pytesseract
pdf2image
Pillow
PyPDF2
https://github.com/tesseract-ocr/tesseract
https://github.com/oschwartz10612/poppler-windows/releases/
